Corbae-Ouliaris

Corbae-Ouliaris (2006) Frequency Domain Filter

Usage

corbae_ouliaris(x, low_freq = NULL, high_freq = NULL)

Arguments

x

Vector, data.frame, matrix or any similar 1D/2D object containing values for filtering.

low_freq

Number indicating the lowest period of oscillation as fractions of \(\pi\). If low_freq > 1, indicating that the direct frequency of the data is provided, this is transformed internally into 2 / high_freq. The default is NULL, meaning that the ifelse(freq > 1, trunc(freq * 1.5), 2) will be used.

high_freq

Number indicating the highest period of oscillation as radians of \(\pi\). If high_freq > 1, indicating that the direct frequency of the data is provided, this is transformed internally into 2 / low_freq. The default is NULL, meaning that the trunc(freq * 8) will be used.

Details

This is a pure R implementation of the filtering algorithm. low_freq and high_freq are connected with characteristics of the series, for example the business circle. low_freq and high_freq must be both either between 0 and 1, meaning that they are frequencies of the period as radians, or both >1, indicating that both are starting and ending periods of the cycle.

low_freq and high_freq are used for keeping the relevant frequencies. These are meant to be the ones inside the range \([ low \_ freq, high \_ freq ]\). Therefore, values outside this range are removed.

For 2-dimensional objects x, filtering per column is applied.

Value

Filtered object with the same length/dimensions and class as the input x.

References

Corbae, D., Ouliaris, S., & Phillips, P. (2002), Band Spectral Regression with Trending-Data. Econometrica 70(3), pp. 1067-1109.

Corbae, D. & Ouliaris, S. (2006), Extracting Cycles from Nonstationary Data, in Corbae D., Durlauf S.N., & Hansen B.E. (eds.). Econometric Theory and Practice: Frontiers of Analysis and Applied Research. Cambridge: Cambridge University Press, pp. 167–177. doi:10.1017/CBO9781139164863.008 .

Shaw, E.S. (1947), Burns and Mitchell on Business Cycles. Journal of Political Economy, 55(4): pp. 281-298. doi:10.1086/256533 .

See also

Examples

# Apply on ts
data(USgdp)
corbae_ouliaris(USgdp, low_freq = 0.0625, high_freq = 0.3333)
#>              Qtr1         Qtr2         Qtr3         Qtr4
#> 1947  220.4374446  122.2146730   33.5540077   -3.6243130
#> 1948   10.6113369   39.3768099   40.2825008   -2.8369519
#> 1949  -71.2495469 -129.0342713 -150.1239076 -133.1706466
#> 1950  -95.7158479  -56.7938430  -24.3861401    3.4074633
#> 1951   28.3929298   45.5817349   46.8861199   31.0777924
#> 1952    9.2724997   -0.8693667   11.4301191   40.0336886
#> 1953   64.0831297   61.5977873   24.7660718  -33.8838686
#> 1954  -88.3356121 -113.8589767 -100.1392564  -55.4793172
#> 1955   -0.9763568   40.9744701   57.5083715   50.4027740
#> 1956   33.4418691   23.4883815   29.8230795   47.7979276
#> 1957   61.2082211   52.8505177   17.0200826  -34.5796837
#> 1958  -78.3242479  -92.5415050  -70.5103180  -23.7320398
#> 1959   26.2614943   60.9672364   73.3421678   66.8071205
#> 1960   48.0627613   21.3612117  -10.9138151  -43.5155312
#> 1961  -66.1001674  -67.8093518  -45.9585810  -11.0506462
#> 1962   17.6993716   24.4018793    7.3201228  -20.0678019
#> 1963  -38.8975425  -38.2269594  -21.8438258   -3.9619638
#> 1964    2.1439936   -6.3598092  -20.7640848  -27.6841375
#> 1965  -17.9379642    8.6409498   44.2687057   77.6716790
#> 1966   98.4785821  100.0390022   81.1337909   46.9447832
#> 1967    8.2971924  -21.9914370  -34.7934942  -29.0828384
#> 1968  -11.7806417    7.0652325   20.0676697   24.7887948
#> 1969   21.3710926    8.8089537  -15.5630829  -52.1882210
#> 1970  -94.8203571 -130.8590427 -147.9745321 -142.3519803
#> 1971 -121.3188483  -97.1623300  -76.5583796  -54.4709513
#> 1972  -18.2375490   40.9211309  118.2813950  195.9186730
#> 1973  252.7403327  276.0212283  265.4426430  227.7381222
#> 1974  168.5070989   89.7860940   -4.3229828 -100.3939317
#> 1975 -175.9900135 -210.4553739 -199.0109947 -158.3363545
#> 1976 -117.4384477  -99.3080375 -106.7697013 -123.0482535
#> 1977 -126.0170019 -104.4491511  -63.7203917  -17.8293211
#> 1978   23.8874393   63.1529127  106.9216434  155.2834856
#> 1979  194.9148847  205.3081392  173.9828761  108.7888798
#> 1980   36.9698356   -9.5064462  -12.6963884   20.8598987
#> 1981   64.1627969   84.7389792   61.0168774   -8.4910677
#> 1982 -106.0221358 -202.6886195 -269.1163515 -284.8615482
#> 1983 -244.5510491 -159.4451321  -53.5437349   45.1688122
#> 1984  115.2085825  148.9431624  153.0223707  141.9249181
#> 1985  128.7384759  118.6677096  108.4987376   91.1764432
#> 1986   61.7702660   21.1277446  -24.2343844  -65.0436569
#> 1987  -93.0069142 -103.7377214  -98.0186802  -81.0750478
#> 1988  -60.0056265  -40.3021758  -23.2197386   -5.7522756
#> 1989   16.3969631   44.4962676   73.9290623   94.8602903
#> 1990   96.2939345   71.3920704   21.4557891  -43.0440108
#> 1991 -104.6902938 -144.1294659 -146.8924011 -110.0714607
#> 1992  -45.6400566   22.7784353   70.1311025   82.8441656
#> 1993   66.7410904   43.8266012   38.5197075   61.5845951
#> 1994  102.5930845  136.4495528  139.6597677  105.0139320
#> 1995   44.7498023  -18.9518492  -68.0997764  -97.2431868
#> 1996 -111.4880538 -118.2637710 -120.9493850 -119.0250548
#> 1997 -112.4454289 -104.2259294  -98.0023526  -93.3269894
#> 1998  -84.3550731  -64.2474745  -31.3853341    9.0860448
#> 1999   50.6088096   92.3518558  140.0824454  197.9927863
#> 2000  258.0455584  297.8499394  292.2432064  231.6136082
#> 2001  132.2367675   28.4674479  -49.0551536  -92.7850223
#> 2002 -119.7077398 -153.5298944 -202.2777062 -249.9755594
#> 2003 -269.3255600 -244.5968520 -185.4001463 -119.9699573
#> 2004  -74.0185611  -53.0208622  -41.9837455  -21.2362280
#> 2005   16.4717582   61.8701246   99.6037011  121.8939344
#> 2006  132.8543243  141.4301653  152.2847921  164.7604551
#> 2007  179.7794946  204.1928338  243.3034318  285.3886795
#> 2008  294.4756212  225.4100350   57.1068615 -179.6092467
#> 2009 -405.3304111 -531.7135006 -513.8078024 -378.2009473
#> 2010 -205.2837944  -77.2522901  -28.9355358  -35.6262331
#> 2011  -43.0597837  -13.0094041   51.1883124  115.4535658
#> 2012  147.9722684  144.3656036  125.5063150  114.1876064
#> 2013  114.4567004  112.2902858   94.7438901   67.3097084
#> 2014   51.6961299   65.5201609  103.0627876  135.7912513
#> 2015  133.3368710   86.8169129   14.1841600  -56.3593536
#> 2016 -109.0939469 -148.9298073 -188.5671686 -228.4239249
#> 2017 -250.1474506 -231.6177808 -170.2218842  -91.6522788
#> 2018  -33.2167413  -13.8549758  -16.8402361   -2.1956084
#> 2019   59.1579559  157.1991746  245.0267087  271.6855323

# Apply on vector
data(USgdp)
corbae_ouliaris(USgdp, low_freq = 0.0625, high_freq = 0.3333)
#>              Qtr1         Qtr2         Qtr3         Qtr4
#> 1947  220.4374446  122.2146730   33.5540077   -3.6243130
#> 1948   10.6113369   39.3768099   40.2825008   -2.8369519
#> 1949  -71.2495469 -129.0342713 -150.1239076 -133.1706466
#> 1950  -95.7158479  -56.7938430  -24.3861401    3.4074633
#> 1951   28.3929298   45.5817349   46.8861199   31.0777924
#> 1952    9.2724997   -0.8693667   11.4301191   40.0336886
#> 1953   64.0831297   61.5977873   24.7660718  -33.8838686
#> 1954  -88.3356121 -113.8589767 -100.1392564  -55.4793172
#> 1955   -0.9763568   40.9744701   57.5083715   50.4027740
#> 1956   33.4418691   23.4883815   29.8230795   47.7979276
#> 1957   61.2082211   52.8505177   17.0200826  -34.5796837
#> 1958  -78.3242479  -92.5415050  -70.5103180  -23.7320398
#> 1959   26.2614943   60.9672364   73.3421678   66.8071205
#> 1960   48.0627613   21.3612117  -10.9138151  -43.5155312
#> 1961  -66.1001674  -67.8093518  -45.9585810  -11.0506462
#> 1962   17.6993716   24.4018793    7.3201228  -20.0678019
#> 1963  -38.8975425  -38.2269594  -21.8438258   -3.9619638
#> 1964    2.1439936   -6.3598092  -20.7640848  -27.6841375
#> 1965  -17.9379642    8.6409498   44.2687057   77.6716790
#> 1966   98.4785821  100.0390022   81.1337909   46.9447832
#> 1967    8.2971924  -21.9914370  -34.7934942  -29.0828384
#> 1968  -11.7806417    7.0652325   20.0676697   24.7887948
#> 1969   21.3710926    8.8089537  -15.5630829  -52.1882210
#> 1970  -94.8203571 -130.8590427 -147.9745321 -142.3519803
#> 1971 -121.3188483  -97.1623300  -76.5583796  -54.4709513
#> 1972  -18.2375490   40.9211309  118.2813950  195.9186730
#> 1973  252.7403327  276.0212283  265.4426430  227.7381222
#> 1974  168.5070989   89.7860940   -4.3229828 -100.3939317
#> 1975 -175.9900135 -210.4553739 -199.0109947 -158.3363545
#> 1976 -117.4384477  -99.3080375 -106.7697013 -123.0482535
#> 1977 -126.0170019 -104.4491511  -63.7203917  -17.8293211
#> 1978   23.8874393   63.1529127  106.9216434  155.2834856
#> 1979  194.9148847  205.3081392  173.9828761  108.7888798
#> 1980   36.9698356   -9.5064462  -12.6963884   20.8598987
#> 1981   64.1627969   84.7389792   61.0168774   -8.4910677
#> 1982 -106.0221358 -202.6886195 -269.1163515 -284.8615482
#> 1983 -244.5510491 -159.4451321  -53.5437349   45.1688122
#> 1984  115.2085825  148.9431624  153.0223707  141.9249181
#> 1985  128.7384759  118.6677096  108.4987376   91.1764432
#> 1986   61.7702660   21.1277446  -24.2343844  -65.0436569
#> 1987  -93.0069142 -103.7377214  -98.0186802  -81.0750478
#> 1988  -60.0056265  -40.3021758  -23.2197386   -5.7522756
#> 1989   16.3969631   44.4962676   73.9290623   94.8602903
#> 1990   96.2939345   71.3920704   21.4557891  -43.0440108
#> 1991 -104.6902938 -144.1294659 -146.8924011 -110.0714607
#> 1992  -45.6400566   22.7784353   70.1311025   82.8441656
#> 1993   66.7410904   43.8266012   38.5197075   61.5845951
#> 1994  102.5930845  136.4495528  139.6597677  105.0139320
#> 1995   44.7498023  -18.9518492  -68.0997764  -97.2431868
#> 1996 -111.4880538 -118.2637710 -120.9493850 -119.0250548
#> 1997 -112.4454289 -104.2259294  -98.0023526  -93.3269894
#> 1998  -84.3550731  -64.2474745  -31.3853341    9.0860448
#> 1999   50.6088096   92.3518558  140.0824454  197.9927863
#> 2000  258.0455584  297.8499394  292.2432064  231.6136082
#> 2001  132.2367675   28.4674479  -49.0551536  -92.7850223
#> 2002 -119.7077398 -153.5298944 -202.2777062 -249.9755594
#> 2003 -269.3255600 -244.5968520 -185.4001463 -119.9699573
#> 2004  -74.0185611  -53.0208622  -41.9837455  -21.2362280
#> 2005   16.4717582   61.8701246   99.6037011  121.8939344
#> 2006  132.8543243  141.4301653  152.2847921  164.7604551
#> 2007  179.7794946  204.1928338  243.3034318  285.3886795
#> 2008  294.4756212  225.4100350   57.1068615 -179.6092467
#> 2009 -405.3304111 -531.7135006 -513.8078024 -378.2009473
#> 2010 -205.2837944  -77.2522901  -28.9355358  -35.6262331
#> 2011  -43.0597837  -13.0094041   51.1883124  115.4535658
#> 2012  147.9722684  144.3656036  125.5063150  114.1876064
#> 2013  114.4567004  112.2902858   94.7438901   67.3097084
#> 2014   51.6961299   65.5201609  103.0627876  135.7912513
#> 2015  133.3368710   86.8169129   14.1841600  -56.3593536
#> 2016 -109.0939469 -148.9298073 -188.5671686 -228.4239249
#> 2017 -250.1474506 -231.6177808 -170.2218842  -91.6522788
#> 2018  -33.2167413  -13.8549758  -16.8402361   -2.1956084
#> 2019   59.1579559  157.1991746  245.0267087  271.6855323

# Apply on matrix per column
mat <- matrix(USgdp, ncol = 4)
corbae_ouliaris(mat, low_freq = 0.0625, high_freq = 0.3333)
#>               [,1]        [,2]        [,3]        [,4]
#>  [1,]  122.7464257 -147.020345 -113.696760   37.143001
#>  [2,]  121.7551863 -139.932119    1.967257   40.795578
#>  [3,]  104.7152492 -114.293150   90.769600  -27.412762
#>  [4,]   81.5982220  -71.118763  117.136438 -129.436200
#>  [5,]   61.8709163  -26.922425   84.482012 -212.278817
#>  [6,]   46.3508669   -1.334270   27.081152 -240.995831
#>  [7,]   26.4847530   -2.007333  -16.204950 -215.775504
#>  [8,]   -7.9317854  -18.813433  -27.252987 -163.813757
#>  [9,]  -58.5287580  -31.379244  -16.333968 -115.494311
#> [10,] -113.8555383  -23.467878   -9.210919  -84.107636
#> [11,] -154.5608891    6.948285  -25.372269  -62.638493
#> [12,] -164.8228428   48.669881  -63.917405  -36.151325
#> [13,] -141.9059944   86.866579 -106.085149    3.233606
#> [14,]  -97.3211436  112.640234 -130.261110   51.207454
#> [15,]  -49.4297712  125.369992 -126.797904   97.845047
#> [16,]  -13.2983564  127.782904 -101.934739  137.343398
#> [17,]    5.3466673  119.950608  -69.813045  171.133805
#> [18,]    9.3484676   98.515848  -40.707453  202.585925
#> [19,]    5.2393092   61.787494  -15.144846  230.499037
#> [20,]   -0.8721754   15.329640   12.691182  249.682787
#> [21,]   -4.1490789  -28.167410   46.951772  259.003197
#> [22,]   -0.2332298  -55.842664   85.227802  268.009799
#> [23,]   13.8865573  -63.103846  120.522836  292.249840
#> [24,]   36.4529842  -56.270748  147.057439  337.027848
#> [25,]   58.1252259  -46.664666  163.556156  381.936837
#> [26,]   64.2936838  -40.632757  170.674362  382.211056
#> [27,]   43.5955521  -33.732015  165.543447  292.440603
#> [28,]   -2.1828904  -14.167901  139.998799  100.124829
#> [29,]  -54.7416957   26.939933   86.259234 -154.713922
#> [30,]  -87.1314957   86.550274    6.761446 -391.516976
#> [31,]  -79.5064621  149.089193  -80.277700 -529.496057
#> [32,]  -32.5861382  194.630779 -146.216097 -532.031300
#> [33,]   31.1695765  209.400045 -166.907926 -426.079690
#> [34,]   80.7080266  190.834461 -137.761091 -283.448041
#> [35,]   95.5993911  144.589502  -78.028333 -176.552137
#> [36,]   78.1982489   77.815620  -20.297530 -138.218034
#> [37,]   50.2264421   -4.336500    8.496081 -150.941342
#> [38,]   36.2641919  -95.934193    2.114567 -169.299916
#> [39,]   46.0058478 -184.965709  -23.541732 -156.323529
#> [40,]   67.8805832 -252.647692  -42.748553 -107.185564
#> [41,]   78.2154655 -280.929159  -37.013339  -45.877869
#> [42,]   58.8637444 -264.078212   -5.560763   -1.842261
#> [43,]   10.4810784 -215.119270   36.420574   12.846723
#> [44,]  -47.5952055 -160.385644   68.670215    8.716272
#> [45,]  -89.8696409 -124.266731   77.475167    5.688492
#> [46,] -100.0613163 -114.408891   60.965905   15.325829
#> [47,]  -78.8459685 -118.189179   26.881811   34.100976
#> [48,]  -40.1706649 -113.181096  -13.257882   51.797589
#> [49,]   -0.8723528  -83.711213  -48.864441   65.304131
#> [50,]   27.9185647  -30.999916  -72.180908   83.850679
#> [51,]   42.3793772   30.305531  -78.867458  120.273681
#> [52,]   42.1361052   85.319965  -69.066524  176.164872
#> [53,]   26.7097846  130.144088  -48.461004  234.883764
#> [54,]   -3.5608573  171.447456  -27.005255  269.969712
#> [55,]  -43.3972746  215.049811  -14.170140  263.105621
#> [56,]  -80.5648310  254.194317  -13.029815  217.156638
#> [57,] -100.0223698  269.323672  -17.863884  153.541411
#> [58,]  -92.7297498  241.864241  -18.194564   96.210923
#> [59,]  -62.6490073  172.086386   -7.027266   55.654067
#> [60,]  -25.7539328   85.940853   13.199906   25.979423
#> [61,]   -0.6125329   22.866437   32.548164   -3.520149
#> [62,]    3.1936438   11.078864   42.743484  -33.776650
#> [63,]   -9.5290274   47.546567   46.133916  -52.557618
#> [64,]  -24.2301054   97.904885   55.761370  -46.431627
#> [65,]  -27.5669052  117.734191   84.301962  -17.785176
#> [66,]  -16.2535889   81.063909  129.347092    9.531940
#> [67,]    2.6139864   -3.536207  167.534007    4.045376
#> [68,]   19.1959279 -100.050172  165.230683  -48.503000
#> [69,]   29.6742816 -169.438647  101.262951 -128.916098
#> [70,]   38.8052922 -193.588870  -13.511428 -192.097278
#> [71,]   54.4390487 -182.445160 -135.269680 -195.991574
#> [72,]   79.0507300 -161.227525 -208.754925 -132.166647
#> [73,]  105.8041713 -149.271491 -199.157048  -35.244676

# Apply on data.frame per column
dfmat <- as.data.frame(mat)
corbae_ouliaris(dfmat, low_freq = 0.0625, high_freq = 0.3333)
#>              V1          V2          V3          V4
#> 1   122.7464257 -147.020345 -113.696760   37.143001
#> 2   121.7551863 -139.932119    1.967257   40.795578
#> 3   104.7152492 -114.293150   90.769600  -27.412762
#> 4    81.5982220  -71.118763  117.136438 -129.436200
#> 5    61.8709163  -26.922425   84.482012 -212.278817
#> 6    46.3508669   -1.334270   27.081152 -240.995831
#> 7    26.4847530   -2.007333  -16.204950 -215.775504
#> 8    -7.9317854  -18.813433  -27.252987 -163.813757
#> 9   -58.5287580  -31.379244  -16.333968 -115.494311
#> 10 -113.8555383  -23.467878   -9.210919  -84.107636
#> 11 -154.5608891    6.948285  -25.372269  -62.638493
#> 12 -164.8228428   48.669881  -63.917405  -36.151325
#> 13 -141.9059944   86.866579 -106.085149    3.233606
#> 14  -97.3211436  112.640234 -130.261110   51.207454
#> 15  -49.4297712  125.369992 -126.797904   97.845047
#> 16  -13.2983564  127.782904 -101.934739  137.343398
#> 17    5.3466673  119.950608  -69.813045  171.133805
#> 18    9.3484676   98.515848  -40.707453  202.585925
#> 19    5.2393092   61.787494  -15.144846  230.499037
#> 20   -0.8721754   15.329640   12.691182  249.682787
#> 21   -4.1490789  -28.167410   46.951772  259.003197
#> 22   -0.2332298  -55.842664   85.227802  268.009799
#> 23   13.8865573  -63.103846  120.522836  292.249840
#> 24   36.4529842  -56.270748  147.057439  337.027848
#> 25   58.1252259  -46.664666  163.556156  381.936837
#> 26   64.2936838  -40.632757  170.674362  382.211056
#> 27   43.5955521  -33.732015  165.543447  292.440603
#> 28   -2.1828904  -14.167901  139.998799  100.124829
#> 29  -54.7416957   26.939933   86.259234 -154.713922
#> 30  -87.1314957   86.550274    6.761446 -391.516976
#> 31  -79.5064621  149.089193  -80.277700 -529.496057
#> 32  -32.5861382  194.630779 -146.216097 -532.031300
#> 33   31.1695765  209.400045 -166.907926 -426.079690
#> 34   80.7080266  190.834461 -137.761091 -283.448041
#> 35   95.5993911  144.589502  -78.028333 -176.552137
#> 36   78.1982489   77.815620  -20.297530 -138.218034
#> 37   50.2264421   -4.336500    8.496081 -150.941342
#> 38   36.2641919  -95.934193    2.114567 -169.299916
#> 39   46.0058478 -184.965709  -23.541732 -156.323529
#> 40   67.8805832 -252.647692  -42.748553 -107.185564
#> 41   78.2154655 -280.929159  -37.013339  -45.877869
#> 42   58.8637444 -264.078212   -5.560763   -1.842261
#> 43   10.4810784 -215.119270   36.420574   12.846723
#> 44  -47.5952055 -160.385644   68.670215    8.716272
#> 45  -89.8696409 -124.266731   77.475167    5.688492
#> 46 -100.0613163 -114.408891   60.965905   15.325829
#> 47  -78.8459685 -118.189179   26.881811   34.100976
#> 48  -40.1706649 -113.181096  -13.257882   51.797589
#> 49   -0.8723528  -83.711213  -48.864441   65.304131
#> 50   27.9185647  -30.999916  -72.180908   83.850679
#> 51   42.3793772   30.305531  -78.867458  120.273681
#> 52   42.1361052   85.319965  -69.066524  176.164872
#> 53   26.7097846  130.144088  -48.461004  234.883764
#> 54   -3.5608573  171.447456  -27.005255  269.969712
#> 55  -43.3972746  215.049811  -14.170140  263.105621
#> 56  -80.5648310  254.194317  -13.029815  217.156638
#> 57 -100.0223698  269.323672  -17.863884  153.541411
#> 58  -92.7297498  241.864241  -18.194564   96.210923
#> 59  -62.6490073  172.086386   -7.027266   55.654067
#> 60  -25.7539328   85.940853   13.199906   25.979423
#> 61   -0.6125329   22.866437   32.548164   -3.520149
#> 62    3.1936438   11.078864   42.743484  -33.776650
#> 63   -9.5290274   47.546567   46.133916  -52.557618
#> 64  -24.2301054   97.904885   55.761370  -46.431627
#> 65  -27.5669052  117.734191   84.301962  -17.785176
#> 66  -16.2535889   81.063909  129.347092    9.531940
#> 67    2.6139864   -3.536207  167.534007    4.045376
#> 68   19.1959279 -100.050172  165.230683  -48.503000
#> 69   29.6742816 -169.438647  101.262951 -128.916098
#> 70   38.8052922 -193.588870  -13.511428 -192.097278
#> 71   54.4390487 -182.445160 -135.269680 -195.991574
#> 72   79.0507300 -161.227525 -208.754925 -132.166647
#> 73  105.8041713 -149.271491 -199.157048  -35.244676